Add search for keys function.
Method searchkeys implements the user interface for search information inside the keys list.
This method is accessed in the main menu in option (2) or by directly using searchKeys().
The prompt ask for a key word to search and prints a list with all the possible items. Once the user choose an item, the interface prints every data of the key.